requirements.txt
python-telegram-bot==21.6
python-dotenv==1.0.0
SQLAlchemy==2.0.0
httpx==0.27
psycopg2-binary==2.9.6
urllib3~=2.2.3
requests~=2.31.0
alembic==1.13.3
click~=8.1.7
validators==0.34.0